+++ Bundle/AdminBundle/src/Command/CreateAdminUserCommand.php
@@ -7,16 +7,16 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\Question;
 use Symfony\Component\Console\Style\SymfonyStyle;
-use Umbrella\AdminBundle\Services\UserManager;
+use Umbrella\AdminBundle\Services\UserManagerInterface;
 
 class CreateAdminUserCommand extends Command
 {
     public const CMD_NAME = 'create:admin_user';
 
-    private UserManager $userManager;
+    private UserManagerInterface $userManager;
     private ?SymfonyStyle $io = null;
 
-    public function __construct(UserManager $userManager)
+    public function __construct(UserManagerInterface $userManager)
     {
         $this->userManager = $userManager;
         parent::__construct(self::CMD_NAME);
@@ -35,7 +35,7 @@ class CreateAdminUserCommand extends Command
      */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $u = $this->userManager->createUser();
+        $u = $this->userManager->create();
 
         $u->firstname = $this->io->askQuestion(new Question('Firstname'));
         $u->lastname = $this->io->askQuestion(new Question('Lastname'));