+++ Bundle/AdminBundle/src/Services/UserMailerInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Umbrella\AdminBundle\Services;
+
+use Umbrella\AdminBundle\Entity\BaseAdminUser;
+
+interface UserMailerInterface
+{
+    public function sendPasswordRequest(BaseAdminUser $user): void;
+}