+++ Bundle/AdminBundle/src/Services/UserManager.php
@@ -3,13 +3,12 @@
 namespace Umbrella\AdminBundle\Services;
 
 use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\ORM\QueryBuilder;
 use Doctrine\Persistence\ObjectRepository;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Umbrella\AdminBundle\Entity\BaseAdminUser;
 use Umbrella\AdminBundle\UmbrellaAdminConfiguration;
 
-class UserManager
+class UserManager implements UserManagerInterface
 {
     protected EntityManagerInterface $em;
     protected UserPasswordHasherInterface $passwordHasher;
@@ -28,42 +27,40 @@ class UserManager
         $this->repo = $this->em->getRepository($this->class);
     }
 
-    public function createUser(): BaseAdminUser
+    public function create(): BaseAdminUser
     {
-        $user = new $this->class();
-
-        return $user;
+        return new $this->class();
     }
 
     public function find($id): ?BaseAdminUser
     {
-        return $this->createQb()
-            ->where('e.id = :id')
-            ->setParameter('id', $id)
-            ->getQuery()
-            ->getOneOrNullResult();
+        return $this->repo->find($id);
     }
 
-    public function findUserByEmail(string $email): ?BaseAdminUser
+    public function findOneBy(array $criteria): ?BaseAdminUser
     {
-        return $this->createQb()
-            ->where('e.email = :email')
-            ->setParameter('email', $email)
-            ->getQuery()
-            ->getOneOrNullResult();
+        return $this->repo->findOneBy($criteria);
     }
 
-    public function findUserByConfirmationToken(string $confirmationToken): ?BaseAdminUser
+    public function findOneByEmail(string $email): ?BaseAdminUser
     {
-        return $this->repo->findOneBy([
-            'confirmationToken' => $confirmationToken,
-        ]);
+        return $this->findOneBy(['email' => $email]);
+    }
+
+    public function findOneByConfirmationToken(string $confirmationToken): ?BaseAdminUser
+    {
+        return $this->findOneBy(['confirmationToken' => $confirmationToken]);
     }
 
     public function updatePassword(BaseAdminUser $user): void
     {
         if (!empty($user->plainPassword)) {
             $user->setPassword($this->passwordHasher->hashPassword($user, $user->plainPassword));
+
+            // erase confirmation token if password was updated
+            $user->passwordRequestedAt = null;
+            $user->confirmationToken = null;
+
             $user->eraseCredentials();
         }
     }
@@ -76,7 +73,7 @@ class UserManager
         $this->em->flush();
     }
 
-    public function remove(BaseAdminUser $user): void
+    public function delete(BaseAdminUser $user): void
     {
         $this->em->remove($user);
         $this->em->flush();
@@ -85,14 +82,5 @@ class UserManager
     public function getClass(): string
     {
         return $this->class;
-    }
-
-    protected function createQb(): QueryBuilder
-    {
-        $qb = $this->em->createQueryBuilder();
-        $qb->select('e');
-        $qb->from($this->class, 'e');
-
-        return $qb;
     }
 }